examples/connector/pyproject.toml
[project]
name = "connector-examples"
version = "0.1.0"
description = "Connector cookbook examples for SmartSearch SDK"
requires-python = ">=3.11"
dependencies = [
    "smart-search-sdk",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
